Formula for upgrading implemented
Only implemented the formula. Fallen behind with all the electricity issues. Need to miss class tomorrow to catch up.

Next up: Complete the interface

+++ Assets/scripts/ClickBehaviour.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class ClickBehaviour : MonoBehaviour
 {
-    public int clickCount;
-    public int clickMultiplier;
+    public float clickCount;
+    public float clickMultiplier;
 
     private void Start()
     {
